--- config/Constants.java
@@ -1,8 +1,0 @@
-package com.example.koifish.config;
-
-public class Constants {
-    public String otpCode() {
-        int code = (int) Math.floor(((Math.random() * 899999) + 100000));
-        return String.valueOf(code);
-    }
-}
