--- config/SecurityConfig.java
@@ -1,121 +1,0 @@
-package com.example.koifish.config;
-
-import com.example.koifish.repository.UserService;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationListener;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.authentication.event.AuthenticationFailureBadCredentialsEvent;
-import org.springframework.security.authentication.event.AuthenticationSuccessEvent;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-
-import java.io.IOException;
-import java.util.HashSet;
-import java.util.Set;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-    @Autowired
-    private UserService userService;
-
-    @Bean
-    public DaoAuthenticationProvider authenticationProvider(){
-        DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
-        authenticationProvider.setUserDetailsService(userService);
-        authenticationProvider.setPasswordEncoder(passwordEncoder());
-        return authenticationProvider;
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authConfig) throws Exception {
-        return authConfig.getAuthenticationManager();
-    }
-
-    @Bean
-    public BCryptPasswordEncoder passwordEncoder(){
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    ApplicationListener<AuthenticationSuccessEvent> successEvent() {
-        return event -> {
-            System.out.println("Success Login " + event.getAuthentication().getClass().getSimpleName() + " - " + event.getAuthentication().getName());
-        };
-    }
-
-    @Bean
-    ApplicationListener<AuthenticationFailureBadCredentialsEvent> failureEvent() {
-        return event -> {
-            System.err.println("Bad Credentials Login " + event.getAuthentication().getClass().getSimpleName() + " - " + event.getAuthentication().getName());
-        };
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .requiresChannel()
-                .requestMatchers(r -> r.getHeader("X-Forwarded-Proto") != null)
-                .requiresSecure()
-                .and()
-                .csrf().disable()
-                .authorizeHttpRequests(authorizeRequests ->
-                        authorizeRequests
-                                .requestMatchers("/", "/home/**", "/index/**", "/movie/**", "/images/**", "/login", "/register", "/oauth/**", "/assets/**", "/auth_assets/**", "/css/**", "/js/**", "/fonts/**", "/img/**", "/sass/**", "/videos/**",
-                                        "/logo.png",    "/1.png", "/2.png", "/3.png", "/4.png", "/5.png", "/6.png", "/slider1.jpg", "/slider2.jpg", "/slider3.jpg")
-                                .permitAll()
-                                .requestMatchers("/client/img/**").permitAll()
-                                .requestMatchers("/client/css/**").permitAll()
-                                .requestMatchers("/client/js/**").permitAll()
-                                .requestMatchers("/chart-data", "/test", "/admin/**", "/userList/**", "/categoryList/**", "/blogTypeList/**")
-                                .hasAnyRole("SUPER_ADMIN", "ADMIN")
-                                .requestMatchers("/client/**", "/profile/**").hasAnyRole("CLIENT")
-                                .anyRequest().authenticated()
-                )
-
-                .formLogin().permitAll()
-                .loginPage("/login")
-                .usernameParameter("username")
-                .passwordParameter("password")
-                .defaultSuccessUrl("/index", true)
-                .failureUrl("/loginError")
-                .and()
-                .logout().logoutSuccessUrl("/").permitAll()
-                .and()
-                .exceptionHandling()
-                .accessDeniedHandler((request, response, accessDeniedException) -> {
-                    response.sendRedirect("/403");
-                })
-                .and()
-                .exceptionHandling()
-                .authenticationEntryPoint((request, response, authException) -> {
-                    response.sendRedirect("/403");
-                })
-                .and()
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.ALWAYS);
-
-        return http.build();
-    }
-
-
-
-}
-
