+++ src/main/java/com/example/koifish/KoiFishApplication.java
@@ -1,0 +1,13 @@
+package com.example.koifish;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KoiFishApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(KoiFishApplication.class, args);
+    }
+
+}
