+++ src/main/java/com/example/koifish/config/Constants.java
@@ -1,0 +1,8 @@
+package com.example.koifish.config;
+
+public class Constants {
+    public String otpCode() {
+        int code = (int) Math.floor(((Math.random() * 899999) + 100000));
+        return String.valueOf(code);
+    }
+}
