+++ src/main/java/com/example/koifish/config/ModelMapperConfig.java
@@ -1,0 +1,13 @@
+package com.example.koifish.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ModelMapperConfig {
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+}
