+++ src/main/java/com/example/koifish/config/SecurityUser.java
@@ -1,0 +1,25 @@
+package com.example.koifish.config;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+
+import java.util.Collection;
+
+public class SecurityUser extends User {
+
+    private static final long serialVersionUID = 1L;
+    private String email;
+
+    public SecurityUser(String username, String password, boolean enabled, boolean accountNonExpired,
+                        boolean credentialsNonExpired, boolean accountNonLocked, Collection<? extends GrantedAuthority> authorities,
+                        String email) {
+        super(username, password, enabled, accountNonExpired, credentialsNonExpired, accountNonLocked, authorities);
+        this.email = email;
+    }
+
+    @Override
+    public String toString() {
+        return "MySecurityUser[email=" + email
+                + "] " + super.toString();
+    }
+}
