+++ src/main/java/com/example/koifish/controller/AuthController.java
@@ -1,0 +1,38 @@
+package com.example.koifish.controller;
+
+
+import com.example.koifish.repository.UserService;
+import jakarta.servlet.http.HttpSession;
+import jakarta.validation.Valid;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+@Controller
+public class AuthController {
+    private final UserService userService;
+    private final PasswordEncoder passwordEncoder;
+
+    public AuthController(UserService userService, PasswordEncoder passwordEncoder) {
+        this.userService = userService;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    @GetMapping("/login")
+    public String showLogin() {
+        return "auth/login";
+    }
+
+    @GetMapping("/logout")
+    public String logoutPage() {
+        SecurityContextHolder.getContext().setAuthentication(null);
+        return "redirect:/login";
+    }
+
+}
