+++ src/main/java/com/example/koifish/controller/FileController.java
@@ -1,0 +1,33 @@
+package com.example.koifish.controller;
+
+import org.springframework.core.io.FileSystemResource;
+import org.springframework.core.io.Resource;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.io.File;
+
+@RestController
+public class FileController {
+
+    private final String UPLOAD_DIR = System.getProperty("user.dir") + File.separator + "uploads"; // Đảm bảo đây là đường dẫn đúng
+
+    // Endpoint phục vụ ảnh hoặc video từ thư mục uploads
+    @GetMapping("/images/{fileName}")
+    public ResponseEntity<Resource> getImage(@PathVariable String fileName) {
+        File file = new File(UPLOAD_DIR + File.separator + fileName);
+
+        if (!file.exists()) {
+            return ResponseEntity.notFound().build();
+        }
+
+        Resource resource = new FileSystemResource(file);
+
+        return ResponseEntity.ok()
+                .header(HttpHeaders.CONTENT_DISPOSITION, "inline; filename=" + fileName)
+                .body(resource);
+    }
+}
