+++ src/main/java/com/example/koifish/model/WaterQuality.class
